+++ packages/terra-form-textarea/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Fixed
+ * Added `size`, `rows`, `disableResize` and `isAutoResizable` props to textarea-field.
+
 ## 5.11.0 - (January 5, 2021)
 
 * Fixed